Complete restructure of the codebase to a more robust, programmatic style
