+++ .gitignore
@@ -1,18 +1,10 @@
 __pycache__/
-.ipynb_checkpoints/
-src/__pycache__/
-dataset/conv/
-dataset/Dusk/
-dataset/fall/
-dataset/model/
-dataset/output_database/
-dataset/output_query/
-dataset/Rain/
-dataset/spring/
-dataset/summer/
-dataset/Sun/
-dataset/winter/
-dataset/event.csv/
-models/VPRTempo78415685001.pth
-models/VPRTempoQuant78415685001.pth
-VPRTempo.egg-info/
+.pyest_cache/
+vprtempo/__pycache__/
+vprtempo/dataset/fall/
+vprtempo/dataset/spring/
+vprtempo/dataset/summer/
+vprtempo/dataset/winter/
+vprtempo/dataset/event.csv
+vprtempo/output/
+vprtempo/src/__pycache__/
