--- dataset/event.csv
@@ -1,160 +1,0 @@
-Image_names,Index
-match_0245.png,0
-match_0247.png,1
-match_0248.png,2
-match_0249.png,3
-match_0250.png,4
-match_0251.png,5
-match_0253.png,6
-match_0256.png,7
-match_0329.png,8
-match_0332.png,9
-match_0333.png,10
-match_0334.png,11
-match_0337.png,12
-match_0338.png,13
-match_0345.png,14
-match_0346.png,15
-match_0347.png,16
-match_0348.png,17
-match_0349.png,18
-match_0350.png,19
-match_0351.png,20
-match_0352.png,21
-match_0354.png,22
-match_0355.png,23
-match_0356.png,24
-match_0357.png,25
-match_0358.png,26
-match_0359.png,27
-match_0361.png,28
-match_0362.png,29
-match_0363.png,30
-match_0364.png,31
-match_0365.png,32
-match_0366.png,33
-match_0367.png,34
-match_0368.png,35
-match_0369.png,36
-match_0370.png,37
-match_0371.png,38
-match_0372.png,39
-match_0373.png,40
-match_0374.png,41
-match_0375.png,42
-match_0376.png,43
-match_0377.png,44
-match_0378.png,45
-match_0379.png,46
-match_0380.png,47
-match_0381.png,48
-match_0382.png,49
-match_0383.png,50
-match_0384.png,51
-match_0385.png,52
-match_0386.png,53
-match_0387.png,54
-match_0388.png,55
-match_0389.png,56
-match_0390.png,57
-match_0391.png,58
-match_0392.png,59
-match_0393.png,60
-match_0394.png,61
-match_0395.png,62
-match_0396.png,63
-match_0397.png,64
-match_0398.png,65
-match_0399.png,66
-match_0400.png,67
-match_0401.png,68
-match_0402.png,69
-match_0403.png,70
-match_0404.png,71
-match_0405.png,72
-match_0491.png,73
-match_0492.png,74
-match_0494.png,75
-match_0538.png,76
-match_0539.png,77
-match_0540.png,78
-match_0541.png,79
-match_0542.png,80
-match_0543.png,81
-match_0544.png,82
-match_0545.png,83
-match_0546.png,84
-match_0547.png,85
-match_0548.png,86
-match_0549.png,87
-match_0550.png,88
-match_0551.png,89
-match_0552.png,90
-match_0553.png,91
-match_0554.png,92
-match_0555.png,93
-match_0720.png,94
-match_0790.png,95
-match_0792.png,96
-match_0793.png,97
-match_0794.png,98
-match_0795.png,99
-match_0796.png,100
-match_0797.png,101
-match_0798.png,102
-match_0799.png,103
-match_0800.png,104
-match_0801.png,105
-match_0802.png,106
-match_0803.png,107
-match_0804.png,108
-match_0805.png,109
-match_0806.png,110
-match_0807.png,111
-match_0808.png,112
-match_0809.png,113
-match_0810.png,114
-match_0811.png,115
-match_0812.png,116
-match_0813.png,117
-match_0814.png,118
-match_0815.png,119
-match_0816.png,120
-match_0817.png,121
-match_0818.png,122
-match_0819.png,123
-match_0820.png,124
-match_0821.png,125
-match_0822.png,126
-match_0823.png,127
-match_0824.png,128
-match_0825.png,129
-match_0826.png,130
-match_0827.png,131
-match_0828.png,132
-match_0829.png,133
-match_0830.png,134
-match_0831.png,135
-match_0832.png,136
-match_0833.png,137
-match_0834.png,138
-match_0835.png,139
-match_0840.png,140
-match_0880.png,141
-match_0881.png,142
-match_0882.png,143
-match_0883.png,144
-match_0884.png,145
-match_0885.png,146
-match_0886.png,147
-match_0887.png,148
-match_0888.png,149
-match_0889.png,150
-match_0890.png,151
-match_0911.png,152
-match_0914.png,153
-match_0919.png,154
-match_0921.png,155
-match_0922.png,156
-match_0924.png,157
-match_0925.png,158
