+++ vprtempo/__init__.py
@@ -1,2 +1,6 @@
-from .VPRTempo import VPRTempo
-from .VPRTempoQuantTrain import VPRTempoQuantTrain
+from . import dataset
+from . import models
+from . import output
+from . import src
+
+__version__ = '1.1.3'
